prop-settings: move styles into react-settings-bridge
